+++ CHANGELOG.md
@@ -11,6 +11,11 @@ All notable changes to this project will be documented in this file.
 * ui/core: Remove everything related to the old deposit return voucher feature
 ### Fixed
 
+## [0.80.3]
+### Added
+* ui: Add listener to `SelfScanningFragment` to react to camera permission changes
+* core: Receive origin type from `PaymentMethodDesricptor` in case it is not set by default
+
 ## [0.80.2]
 ### Fixed
 * ui: Add missing focus to the product search field after opening the `ProductSearchView`